third_party/blink/web_tests/platform/generic/fast/loader/stateobjects/state-attribute-popstate-event-expected.txt

Check that PopStateEvent.state always has a correct value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS history.state is defined.
Push state 1
Push state 2
PASS history.state is 2
Go back
PASS popStateEvent.state is 1
PASS history.state is 1
Push state 3
PASS popStateEvent.state is 1
PASS history.state is 3
PASS successfullyParsed is true

TEST COMPLETE
